saloonbackend/requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
flask-cors
